refactor: move components into separate folders
