+++ src/components/App/App.tsx
@@ -1,0 +1,11 @@
+import NavBar from "../NavBar/NavBar";
+import Main from "../Main/Main";
+
+export default function App() {
+  return (
+    <div className="h-svh bg-slate-900 p-6 text-slate-100">
+      <NavBar />
+      <Main />
+    </div>
+  );
+}
