+++ src/components/NavBar/Logo.tsx
@@ -1,0 +1,10 @@
+function Logo() {
+  return (
+    <div className="flex items-center gap-2">
+      <span role="img">🍿</span>
+      <h1 className="text-2xl font-semibold">usePopcorn</h1>
+    </div>
+  );
+}
+
+export default Logo;
