+++ src/components/NavBar/NumResults.tsx
@@ -1,0 +1,5 @@
+function NumResults() {
+  return <p className="text- justify-self-end text-lg">Found X results</p>;
+}
+
+export default NumResults;
