+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App";
+import App from "./components/App/App";
 import "./index.css";
 
 const rootElement = document.getElementById("root");
